requirements.txt
aniso8601==10.0.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.0.3
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.31
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
